Move stuff to shared
